Semester Project README.md
Phonebook Application
The Phonebook Application is made for storing and organizing contact 
information for people and businesses. Users can manage their contact list and 
save and search for contact information using this project. This project will offer 
a lot of functions, including name or phone number searching and sorting. Users 
can easily save and manage contact information with the C++ phonebook 
application. It provides users with an easy way to organize and retain contact 
information. The system is additionally created to be secure, ensuring that all 
user data is kept private